linux_resource_monitor/requirements.txt
# ========================
# requirements.txt
# ========================
psutil
PyYAML
fastapi
uvicorn
dash
plotly
pytest
requests
influxdb-client==1.49.0
